♻️ Refactor: 그룹 초대 페이지 로직 분리
