+++ src/pages/group/GroupInviteReceivePage.tsx
@@ -1,30 +1,11 @@
 import Button from '@/components/common/button/Button/Button';
 import Header from '@/components/common/header/Header';
 import InputBox from '@/components/common/input/InputBox';
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { postJoinGroup } from '@/services/group/postJoinGroup';
-import { useToast } from '@/hooks/use-toast';
 import MetaTags from '@/components/common/metaTags/MetaTags';
+import { useGroupInvite } from '@/hooks/useGroupInvite';
 
 const GroupInviteReceivePage = () => {
-  const [inviteLink, setInviteLink] = useState('');
-  const navigate = useNavigate();
-  const { toast } = useToast();
-
-  const handleBack = () => {
-    navigate('/group-select');
-  };
-
-  const handleGoIn = async () => {
-    try {
-      const joinResult = await postJoinGroup({ inviteLink });
-      navigate(`/main/${joinResult.result.channelId}`);
-    } catch (error) {
-      toast({ title: '코드를 다시 확인해주세요' });
-      console.error('그룹 입장 실패:', error);
-    }
-  };
+  const { inviteLink, setInviteLink, handleBack, handleGoIn } = useGroupInvite();
 
   return (
     <div className={`flex h-screen flex-col`}>
